+++ CHANGELOG.md
@@ -3,6 +3,14 @@ Tidarr notable changes.
 
 [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) format.
 
+## 📦 1.1.1
+### 🚀 Added
+- [Front] Add API key manager in configuration dialog
+- [API] Add X-API-KEY header to authenticate external queries
+- [API] Add replay gain post processing step (REPLAY_GAIN=true)
+- [API] BETA - Add Lidarr NZB indexer endpoints
+- [API] BETA - Add Lidarr Sabnzbd download client endpoints
+
 ## 📦 1.1.0
 ### 🚀 Added
 - [API] Add support for OIDC authentication #409